pyproject.toml
[project]
name = "intro-langgraph"
version = "0.1.0"
description = "LangGraph Academia: Introduction to LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph[graphviz]",
    "langgraph-prebuilt",
    "langgraph-sdk",
    "langgraph-checkpoint-sqlite",
    "langsmith",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "tavily-python",
    "wikipedia",
    "trustcall",
    "langgraph-cli[inmem]",
    "python-dotenv>=1.1.1",
    "langchain-tavily>=0.2.11",
    "langgraph-checkpoint-postgres>=2.0.23",
    "psycopg2-binary>=2.9.10",
    "langfuse>=3.3.0",
]
